Create text-list.php
Hi, 
The idea of this fork is to create metabox support for a new type of clonable field: grouped text boxes. I'm thinking it would create an array of arrays like checkbox-list, but I'm having trouble getting it supported in helpers.php. Php insight and contributions would be much appreciated.
 
Example of desired result

________(name) _________(favorite color) 
________(name) _________(favorite color) 
________(name) _________(favorite color)   (+) clone