Fixed bug that keeps new cloned fields hidden
This commit fixes a bug that exists when trying to add a new cloned select2 dropdown. The element stored in `$this` had `style="display: none;` added to it when it was originally "selectized" on instantiation, after clone this attribute was being retained thus hiding the new select2 container. The width of the cloned select2 container was also off so I also added code to mimic the existing select2s.