Remove reset function for cloning
Resetting all names and ids for all clone fields in a set breaks jQuery
fields like `date` when `sort_clone` is set to true, which rely on an id
to work properly.  Instead, each new clone is given an incremental index
number.  This works with regular cloning as well as sorting.