Prioritize using field storage to get field value.
Previously, to get field value, the plugin checks object_type first,
then storage, then fallback to post storage. But since the storage is
attached to fields when a meta box is created, we're safe to use it.

For old versions of plugins, the storage is null, which won't be
attached to fields.

This commit reverts the commit:
https://github.com/wpmetabox/meta-box/commit/75ca5078c8b400e6e9909eee74e99e28e6ebac9f