Fixed taxonomy field not showing correct saved value
The taxonomy field doesn't store value in the custom fields, so
metadata_exists() doesn't work for it. Thus, is_saved() is false and the
field uses the 'std' value instead (which is empty).

Sample code and issue: https://wordpress.org/support/topic/doesnt-save-taxonomy-value/

This commit remove the use of metadata_exists() and use the old method
to check if value exists. Also reverts PR #1268.