Fixed reorder images delete images when force_delete=true
The problem is when reorder images, the collection tries to set order
for models by removing and readding them. The remove action triggers an
ajax request by WordPress (in WordPress media.js file) that remove the
file.

Solution: remove the code to set order for models. It's not needed.

Bug is reported here: https://metabox.io/support/topic/image-upload-image-is-deleted-after-change-position-using-drag-and-drop/