Fixed cannot save posts when select_tree has required attribute.
Link: https://metabox.io/support/topic/error-in-select-tree-taxonomy/

The taxonomy has some term nested in 3 levels, but it also has other nested terms in only 2 levels. Then the validation of the abligatory field fails because apparently it is waiting for a third level, even though it should not be like that.

This fix set/unset the 'required' attribute correspondingly when select
is visible/hidden (similar to Conditional Logic). So doesn't affect the
valiation when saving.