Google Maps & OSM: ask for current user location for the map (#1537)
This PR will ask users for permission to access the current location via Geolocation API, which is available in all browsers and use that location to set the current map center and for the marker.

Previously, the plugin always loads the Dublin location as default, which causes a confusion for users in other countries.

This "ask" won't happen if there's a saved location, or a default location. In other words, it only asks if there's no location is saved or set as default.